# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AuthResponse" : {
        "properties" : {
          "accessToken" : {
            "example" : "eyJhbGciOiJIUzI1NiJ9...",
            "type" : "string"
          },
          "expiresIn" : {
            "example" : 900,
            "format" : "int64",
            "type" : "integer"
          },
          "refreshToken" : {
            "example" : "0G2f...",
            "type" : "string"
          },
          "tokenType" : {
            "example" : "Bearer",
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/SessionUserDto"
          }
        },
        "type" : "object"
      },
...
      "ErrorResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "details" : {
            "items" : {
              "$ref" : "#/components/schemas/FieldIssue"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "FieldIssue" : {
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "issue" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RefreshRequest" : {
        "properties" : {
          "refreshToken" : {
            "example" : "0G2f...",
            "maxLength" : 512,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "refreshToken" ],
        "type" : "object"
      },
...
      "SessionUserDto" : {
        "properties" : {
          "displayName" : {
            "example" : "Alice Analyst",
            "type" : "string"
          },
          "permissions" : {
            "example" : [ "ioc:read", "ioc:submit" ],
            "items" : {
              "type" : "string"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "role" : {
            "example" : "TENANT_ADMIN",
            "type" : "string"
          },
          "tenantId" : {
            "example" : "8b1a9c33-2f4e-4d55-9f6a-0c1d2e3f4a5b",
            "type" : "string"
          },
          "userId" : {
            "example" : "3f1b0c2e-9a4d-4c1a-8b77-2b0f1a9c5d10",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "apiKeyAuth" : {
        "description" : "Machine-to-machine key: ctip_<env>_<32 base62>",
        "in" : "header",
        "name" : "X-API-Key",
        "type" : "apiKey"
      },
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from POST /api/v1/auth/login",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Cyber Threat Intelligence Platform read API. M1 exposes anonymous read access to public TLP:CLEAR intelligence with cursor pagination; authentication, RBAC and write endpoints arrive in M2. All errors share a single ErrorResponse shape with a log-correlated traceId (see docs/spec/09-api.md §9.4).",
    "title" : "CTIP API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/auth/refresh" : {
      "post" : {
        "description" : "Exchanges a refresh token for a new pair. The presented token is consumed; replaying an already-used token revokes the whole token family. 認證:以請求主體中的 refresh token 認證,不需 Authorization 標頭。",
        "operationId" : "refresh",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : {
                  "accessToken" : "eyJhbGciOiJIUzI1NiJ9.e30.sig",
                  "expiresIn" : 900,
                  "refreshToken" : "0G2f8xQ...",
                  "tokenType" : "Bearer",
                  "user" : {
                    "displayName" : "Alice Analyst",
                    "permissions" : [ "ioc:read", "ioc:submit" ],
                    "role" : "TENANT_ADMIN",
                    "tenantId" : "8b1a9c33-2f4e-4d55-9f6a-0c1d2e3f4a5b",
                    "userId" : "3f1b0c2e-9a4d-4c1a-8b77-2b0f1a9c5d10"
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "New session issued"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "Unknown, revoked, expired or reused token (UNAUTHENTICATED)"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit exceeded (RATE_LIMIT_EXCEEDED)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected error (INTERNAL_ERROR)"
          }
        },
        "security" : [ ],
        "summary" : "Rotate a refresh token",
        "tags" : [ "Auth" ]
      }
    },
...
    "/api/v1/stix/bundle" : {
      "get" : {
        "description" : "Exports visible, redistributable indicators plus the referenced TLP 2.0 marking-definitions as one STIX 2.1 bundle. Object count is capped by plan (PLAN_LIMIT_EXCEEDED above the cap). 認證:stix:export——匿名不可用;M1 尚無登入流程,一律回 403,Phase 13 起依 RBAC 判定。",
        "operationId" : "bundle",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : {
                  "id" : "bundle--7c94...",
                  "objects" : [ {
                    "type" : "marking-definition"
                  }, {
                    "type" : "indicator"
                  } ],
                  "type" : "bundle"
                }
              }
            },
            "description" : "STIX 2.1 bundle"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "FORBIDDEN (anonymous) / PLAN_LIMIT_EXCEEDED (over object cap)"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit exceeded (RATE_LIMIT_EXCEEDED)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected error (INTERNAL_ERROR)"
          }
        },
        "security" : [ ],
        "summary" : "Export a STIX 2.1 bundle",
        "tags" : [ "STIX" ]
      }
    },
    "/api/v1/stix/{stixId}" : {
      "get" : {
        "description" : "Returns a stored STIX 2.1 object by its STIX id: indicator projections (indicator--{uuid}) after visibility and redistribution filtering, and the five fixed TLP 2.0 marking-definitions served from constants. 認證:匿名。",
        "operationId" : "byStixId",
        "parameters" : [ {
          "description" : "STIX id, e.g. indicator--{uuid} or marking-definition--{uuid}",
          "in" : "path",
          "name" : "stixId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : {
                  "created" : "2022-10-01T00:00:00.000Z",
                  "id" : "marking-definition--94868c89-83c2-464b-929b-a1a8aa3c8487",
                  "name" : "TLP:CLEAR",
                  "spec_version" : "2.1",
                  "type" : "marking-definition"
                }
              }
            },
            "description" : "The STIX object (verbatim stored JSON for indicators)"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "NOT_FOUND (unknown id, invisible or non-redistributable)"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit exceeded (RATE_LIMIT_EXCEEDED)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected error (INTERNAL_ERROR)"
          }
        },
        "security" : [ ],
        "summary" : "Get one STIX object",
        "tags" : [ "STIX" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ],
...
}